requirements.txt
flask
flask-cors
web3
py-solc-x
pymongo
python-dotenv
pytest
